+++ packages/demos/kitchen-sink/stories/helpers/data.ts
@@ -2,6 +2,7 @@
 // Copyright 2022 DXOS.org
 //
 
+import faker from 'faker';
 import { useEffect, useMemo, useState } from 'react';
 
 import { Party } from '@dxos/client';
@@ -70,10 +71,12 @@ export const useGraphModel = (party?: Party): EchoGraphModel => {
   return model;
 };
 
+type TestPartyCallback = (builder: PartyBuilder) => Promise<void>;
+
 /**
  * Generate test party.
  */
-export const useTestParty = (): Party | undefined => {
+export const useTestParty = (callback: TestPartyCallback = buildTestParty): Party | undefined => {
   const client = useClient();
   const [party, setParty] = useState<Party>();
   const builder = usePartyBuilder(party);
@@ -81,6 +84,7 @@ export const useTestParty = (): Party | undefined => {
   useEffect(() => {
     setImmediate(async () => {
       const party = await client.echo.createParty();
+      await party.setTitle(faker.lorem.word());
       setParty(party);
     });
   }, []);
@@ -88,7 +92,7 @@ export const useTestParty = (): Party | undefined => {
   useEffect(() => {
     if (builder) {
       setImmediate(async () => {
-        await buildTestParty(builder);
+        await callback(builder);
       }, []);
     }
   }, [builder, party]);
@@ -100,7 +104,7 @@ export const useTestParty = (): Party | undefined => {
  * Build the party.
  * @param builder
  */
-export const buildTestParty = async (builder: PartyBuilder) => {
+export const buildTestParty: TestPartyCallback = async (builder: PartyBuilder) => {
   await builder.createOrgs([3, 7], async (orgBuilder: OrgBuilder) => {
     await orgBuilder.createPeople([3, 10]);
     await orgBuilder.createProjects([2, 7], async (projectBuilder: ProjectBuilder) => {