+++ packages/sdk/react-framework/src/components/index.ts
@@ -8,4 +8,5 @@ export * from './ExportDialog';
 export * from './ImportDialog';
 export * from './ProfileDialog';
 export * from './RegistrationDialog';
+export * from './TestInvitationDialog';
 export * from './Theme';