+++ CHANGELOG.md
@@ -1,5 +1,47 @@
 # Change Log
 
+### [2.28.10](https://www.github.com/dxos/protocols/compare/v2.28.9...v2.28.10) (2022-03-23)
+
+
+### Bug Fixes
+
+* SharingDialog message & registry client config ([#1077](https://www.github.com/dxos/protocols/issues/1077)) ([ec8ff7b](https://www.github.com/dxos/protocols/commit/ec8ff7bcf557fb37730dcb6a69512b430deb74fb))
+
+### [2.28.9](https://www.github.com/dxos/protocols/compare/v2.28.8...v2.28.9) (2022-03-22)
+
+
+### Features
+
+* Kitchen sink export party  ([#1073](https://www.github.com/dxos/protocols/issues/1073)) ([0b19adb](https://www.github.com/dxos/protocols/commit/0b19adb4ed26f981d9effc7f6d3f414c4010b588))
+
+### [2.28.8](https://www.github.com/dxos/protocols/compare/v2.28.7...v2.28.8) (2022-03-22)
+
+
+### Features
+
+* **kitchen-sink:** Add party import export. ([#1060](https://www.github.com/dxos/protocols/issues/1060)) ([54f6fa7](https://www.github.com/dxos/protocols/commit/54f6fa7b40735cfe4199b2c0ea4c00f16303e5d2))
+
+
+### Bug Fixes
+
+* Config loading ([#1071](https://www.github.com/dxos/protocols/issues/1071)) ([8f80020](https://www.github.com/dxos/protocols/commit/8f800207e0024e379a7683d51f16b7022b85a74f))
+
+### [2.28.7](https://www.github.com/dxos/protocols/compare/v2.28.6...v2.28.7) (2022-03-21)
+
+
+### Bug Fixes
+
+* Added links in snapshot creation ([#1061](https://www.github.com/dxos/protocols/issues/1061)) ([7ae4894](https://www.github.com/dxos/protocols/commit/7ae4894371bbe402b0e4eceef35ef0f98c46c20e))
+* move running bots out of cwd & ipfs fix. ([#1069](https://www.github.com/dxos/protocols/issues/1069)) ([1dc18be](https://www.github.com/dxos/protocols/commit/1dc18bee136b00f6a728119ed290cdd98758774c))
+
+### [2.28.6](https://www.github.com/dxos/protocols/compare/v2.28.5...v2.28.6) (2022-03-19)
+
+
+### Features
+
+* file upload/download. ([#1063](https://www.github.com/dxos/protocols/issues/1063)) ([73979ec](https://www.github.com/dxos/protocols/commit/73979ec45f86584aa4d617a887b9f179c7660842))
+* Sharing dialog with bot invitations. ([#1058](https://www.github.com/dxos/protocols/issues/1058)) ([50c5d54](https://www.github.com/dxos/protocols/commit/50c5d54b459f1c31ff1c78fd61e8c99ad7c874c7))
+
 ### [2.28.5](https://www.github.com/dxos/protocols/compare/v2.28.4...v2.28.5) (2022-03-16)